solution/requirements-dev.txt
autopep8==1.4.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.8
djangorestframework==3.11.0
pep8==1.7.1
pycodestyle==2.5.0
python-decouple==3.3
pytz==2018.5
sqlparse==0.2.4
static3==0.7.0
